extract common functions in version finder to base_version_finder
