+++ maven/lib/dependabot/maven/update_checker/version_finder.rb
@@ -6,13 +6,13 @@ require "dependabot/package/release_cooldown_options"
 require "dependabot/update_checkers/version_filters"
 require "dependabot/maven/package/package_details_fetcher"
 require "dependabot/maven/update_checker"
-require "dependabot/maven/shared/shared_version_finder"
+require "dependabot/maven/shared/base_version_finder"
 require "sorbet-runtime"
 
 module Dependabot
   module Maven
     class UpdateChecker
-      class VersionFinder < Dependabot::Maven::Shared::SharedVersionFinder
+      class VersionFinder < Dependabot::Maven::Shared::BaseVersionFinder
         extend T::Sig
 
         sig do
@@ -52,91 +52,12 @@ module Dependabot
           @package_details ||= package_details_fetcher.fetch
         end
 
-        sig { returns(T::Array[Dependabot::Package::PackageRelease]) }
-        def releases
-          (package_details&.releases || []).reverse
-        end
-
-        sig { returns(T.nilable(T::Hash[T.untyped, T.untyped])) }
-        def latest_version_details
-          release = fetch_latest_release
-          release&.version ? { version: release.version, source_url: release.url } : nil
-        end
-
-        sig { returns(T.nilable(T::Hash[T.untyped, T.untyped])) }
-        def lowest_security_fix_version_details
-          release = fetch_lowest_security_fix_release
-          release&.version ? { version: release.version, source_url: release.url } : nil
-        end
-
-        protected
-
-        sig { returns(T::Boolean) }
-        def cooldown_enabled?
-          true
-        end
-
-        sig do
-          params(language_version: T.nilable(T.any(String, Dependabot::Version)))
-            .returns(T.nilable(Dependabot::Version))
-        end
-        def fetch_latest_version(language_version: nil)
-          fetch_latest_release(language_version: language_version)&.version
-        end
-
-        sig do
-          params(language_version: T.nilable(T.any(String, Dependabot::Version)))
-            .returns(T.nilable(Dependabot::Version))
-        end
-        def fetch_latest_version_with_no_unlock(language_version:)
-          fetch_latest_release(language_version: language_version)&.version
-        end
-
-        sig do
-          params(language_version: T.nilable(T.any(String, Dependabot::Version)))
-            .returns(T.nilable(Dependabot::Version))
-        end
-        def fetch_lowest_security_fix_version(language_version: nil)
-          fetch_lowest_security_fix_release(language_version: language_version)&.version
-        end
-
-        sig do
-          params(language_version: T.nilable(T.any(String, Dependabot::Version)))
-            .returns(T.nilable(Dependabot::Package::PackageRelease))
-        end
-        def fetch_latest_release(language_version: nil) # rubocop:disable Lint/UnusedMethodArgument
-          possible_releases = filter_prerelease_versions(releases)
-          possible_releases = filter_date_based_versions(possible_releases)
-          possible_releases = filter_version_types(possible_releases)
-          possible_releases = filter_ignored_versions(possible_releases)
-          possible_releases = filter_by_cooldown(possible_releases)
-          possible_releases_reverse = possible_releases.reverse
-
-          possible_releases_reverse.find do |r|
-            package_details_fetcher.released?(r.version)
-          end
-        end
-
-        sig do
-          params(language_version: T.nilable(T.any(String, Dependabot::Version)))
-            .returns(T.nilable(Dependabot::Package::PackageRelease))
-        end
-        def fetch_lowest_security_fix_release(language_version: nil) # rubocop:disable Lint/UnusedMethodArgument
-          possible_releases = filter_prerelease_versions(releases)
-          possible_releases = filter_date_based_versions(possible_releases)
-          possible_releases = filter_version_types(possible_releases)
-          possible_releases = Dependabot::UpdateCheckers::VersionFilters
-                              .filter_vulnerable_versions(
-                                possible_releases,
-                                security_advisories
-                              )
-          possible_releases = filter_ignored_versions(possible_releases)
-          possible_releases = filter_lower_versions(possible_releases)
-
-          possible_releases.find { |r| package_details_fetcher.released?(r.version) }
-        end
-
         private
+
+        sig { override.params(version: Dependabot::Version).returns(T::Boolean) }
+        def released?(version)
+          package_details_fetcher.released?(version)
+        end
 
         sig { returns(Package::PackageDetailsFetcher) }
         def package_details_fetcher
